Fixed the Jumping Animation
I fixed the jumping animation. It looks like it will be easier to
control all the physics through, so if we keep all the animations in
place (no movement in the x and y planes just animate the jumping like
he doesn't move) it will be smoother adding the animations later. I also
had to remove a few of the frames at the end, we can add them back in it
looks goofy