Added Sound Effects, Disabled Fist Colliders
I added sound effects and background music to our scenes.

Colton - I modified the GUIButton and Map classes to implement audio.
Each button that is created will need an audio component added to it
with the sound it should play when it's selected

The colliders on fists are disabled when not punching to avoid
triggering damage/recoil when not punching. Right now, punching while
crouched causes a punch to trigger still without changing the animation.
Once we get an animation in this will be fixed.