Added the last of the Intros/Outros
So the "You Win" Button text has been replaced by some other text.
(Don't know if I accidently did that, but I'm not sure how to fix it).
Also, noticed a little mishap in the names and placement.  Fight two is
supposed to be against Benedict Arnold, but has the image of Cornwallis,
so I added cornwallis's text fight (I didn't see the mislabeled map
screen until after)  And you fight Arnold at the battle of Yorktown, so
either the names need to be fixed, or the text and the portraits need to
be swapped.  I think Cornwallis should be the third fight.